// Costa_Blanca-Front/src/app/interfaces/dados-empresa.ts
export interface IDadosEmpresaInterface {
  bairroEscritorio: string;
  bairroLoja: string;
  cepEscritorio: string;
  cepLoja: string;
  cidadeEscritorio: string;
  cidadeLoja: string;
  cnpj: string;
  complementoEscritorio: string;
  complementoLoja: string;
  cpfResponsavel: string;
  email: string;
  emailResponsavel: string;
  estadoEscritorio: string;
  estadoLoja: string;
  id: number;
  inscricaoEstadual: string;
  inscricaoMunicipal: string;
  logradouroEscritorio: string;
  logradouroLoja: string;
  nomeFantasia: string;
  nomeResponsavel: string;
  numeroEscritorio: string;
  numeroLoja: string;
  razaoSocial: string;
  site: string;
  telefone: string;
  telefone2: string;
  telefoneResponsavel: string;
}